requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
sqlalchemy
boto3
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
pytest
httpx 